+++ packages/react-editor/src/website/Root.e2e.js
@@ -208,11 +208,4 @@ async function testCreateNewForm(page) {
   // click remove form
 
   // verify form is not in the list anymore
-
-
-
-
-
-
-
 }