+++ CHANGELOG.md
@@ -14,6 +14,10 @@ Changes that have landed in master but are not yet released
 
 * Add package `react-editor` - demo website to create form configuration using a simple UI
 
+### Documentation
+
+* Add online form editor of `react-editor` package to the docs
+
 ## v1.0.4 (February 24, 2020)
 
 ### React Kit